Added a trailing / on WHELK_HOST and updated fabfile for demo deploy.

+++ fabfile.py
@@ -7,17 +7,24 @@ from fabric.api import *
 from sqlalchemy import *
 from sqlalchemy.orm import *
 
-#env.hosts = ['192.168.3.128']
-env.hosts = ['devlab.libris.kb.se']
-env.user = 'jenkins'
-wwwuser = 'apache'
-wwwgroup = 'apache'
-#virtualenv_home = '/srv/data/virtualenvs/505'
 kitin_path = '/srv/www/kitin'
-#kitin_path = '/Library/WebServer/kitin'
 
 cfg = {}
 #execfile(os.path.join(os.path.dirname(__file__), 'config.cfg'), cfg)
+
+@task
+def lab():
+    env.hosts = ['devlab.libris.kb.se']
+    env.user = 'jenkins'
+    env.wwwuser = 'apache'
+    env.wwwgroup = 'apache'
+
+@task
+def demo():
+    env.hosts = ['193.10.75.247']
+    env.user = 'riakcluster4'
+    env.wwwuser = '_www'
+    env.wwwgroup = '_www'
 
 @task
 def create_db():
@@ -62,21 +69,24 @@ def fetch_vendor_assets():
 
 @task
 def deploy():
-    prepare()
-    sudo('rm -f /tmp/kitin.tgz')
-    local('tar cfz /tmp/kitin.tgz --exclude=\'.*\' *')
-    put('/tmp/kitin.tgz', '/tmp/')
-    sudo('rm -fr %s' % kitin_path)
-    sudo('mkdir -m 775 %s' % kitin_path)
-    sudo('chown %s:%s %s' % (env.user, wwwgroup, kitin_path))
-    run('mkvirtualenv kitin')
-    with cd('%s' % kitin_path):
-        run('tar xzf /tmp/kitin.tgz')
-        run('python tools/create_wsgi_file.py')
-        sudo('chown %s kitin.db' % wwwuser)
+    if not env.host:
+        print "Call lab or demo first, i.e: $>fab lab deploy"
+    else:
+        prepare()
+        sudo('rm -f /tmp/kitin.tgz')
+        local('tar cfz /tmp/kitin.tgz --exclude=\'.*\' *')
+        put('/tmp/kitin.tgz', '/tmp/')
+        sudo('rm -fr %s' % kitin_path)
+        sudo('mkdir -m 775 %s' % kitin_path)
+        sudo('chown %s:%s %s' % (env.user, env.wwwgroup, kitin_path))
+        run('mkvirtualenv kitin')
+        with cd('%s' % kitin_path):
+            run('tar xzf /tmp/kitin.tgz')
+            run('python tools/create_wsgi_file.py')
+            sudo('chown %s kitin.db' % env.wwwuser)
 
-    with prefix('workon kitin'):
-        run('pip install -r %s/dev-requirements.txt' % kitin_path)
+        with prefix('workon kitin'):
+            run('pip install -r %s/dev-requirements.txt' % kitin_path)
 
 @task
 def clear_config():
